requirements.txt
# FastAPI Backend Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# HTTP & Async Client
aiohttp==3.9.1
httpx==0.25.2

# PDF Processing
PyMuPDF==1.23.8
PyPDF2==3.0.1
python-docx==1.1.0

# Vector Database & Embeddings
pinecone-client==2.2.4
faiss-cpu==1.7.4
sentence-transformers==2.2.2

# OpenAI & LLM
openai==1.99.2
tiktoken==0.10.0

# LangChain for LLM orchestration, JSON parsing, and parallel processing
langchain==0.3.27
langchain-openai==0.3.28
langchain-core==0.3.74
langchain-text-splitters==0.3.9

# Data Processing
pandas==2.1.4
numpy==1.24.4
python-dotenv==1.0.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Logging & Monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Database
motor==3.3.2
pymongo==4.6.0

# Text Processing
spacy==3.7.2
nltk==3.8.1

# Utilities
pydantic==2.5.0
python-dateutil==2.8.2
requests==2.31.0
